Update surv_neyman.c
While running this package in RSudio I've run into the following issue: "twosample_neyman" not available for .C() for package "surv2sample", which after a great deal of research seems to be an issue of the functions not being exported. Using Rccp this should export the function to then be called by surv2.neyman.